# scripts/sdk/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "nocodb",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
...
    "/api/v1/db/data/{orgs}/{projectName}/{tableName}": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "orgs",
          "in": "path",
          "required": true
        },
        {
          "schema": {
            "type": "string"
          },
          "name": "projectName",
          "in": "path",
          "required": true
        },
        {
          "schema": {
            "type": "string"
          },
          "name": "tableName",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "summary": "Table row list",
        "operationId": "db-table-row-list",
        "description": "",
        "tags": [
          "DB table row"
        ],
        "parameters": [
          {
            "schema": {
              "type": "array"
            },
            "in": "query",
            "name": "fields"
          },
          {
            "schema": {
              "type": "array"
            },
            "in": "query",
            "name": "sort"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "where"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "offset"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "limit"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "summary": "Table row create",
        "operationId": "db-table-row-create",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "tags": [
          "DB table row"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          }
        }
      }
    },
    "/api/v1/db/data/{orgs}/{projectName}/{tableName}/find-one": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "orgs",
          "in": "path",
          "required": true
        },
        {
          "schema": {
            "type": "string"
          },
          "name": "projectName",
          "in": "path",
          "required": true
        },
        {
          "schema": {
            "type": "string"
          },
          "name": "tableName",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "summary": "Table row FindOne",
        "operationId": "db-table-row-find-one",
        "description": "",
        "tags": [
          "DB table row"
        ],
        "parameters": [
          {
            "schema": {
              "type": "array"
            },
            "in": "query",
            "name": "fields"
          },
          {
            "schema": {
              "type": "array"
            },
            "in": "query",
            "name": "sort"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "where"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/v1/db/data/{orgs}/{projectName}/{tableName}/groupby": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "orgs",
          "in": "path",
          "required": true
        },
        {
          "schema": {
            "type": "string"
          },
          "name": "projectName",
          "in": "path",
          "required": true
        },
        {
          "schema": {
            "type": "string"
          },
          "name": "tableName",
          "in": "path",
          "required": true
        },
        {
          "schema": {
            "type": "string"
          },
          "in": "query",
          "name": "column_name",
          "description": "Column name of the column you want to group by, eg. `column_name=column1`"
        }
      ],
      "get": {
        "summary": "Table row Group by",
        "operationId": "db-table-row-group-by",
        "description": "",
        "tags": [
          "DB table row"
        ],
        "parameters": [
          {
            "schema": {
              "type": "array"
            },
            "in": "query",
            "name": "sort"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "where"
          },
          {
            "schema": {
              "type": "number"
            },
            "in": "query",
            "name": "limit"
          },
          {
            "schema": {
              "type": "number"
            },
            "in": "query",
            "name": "offset"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/v1/db/data/{orgs}/{projectName}/{tableName}/views/{viewName}/group/{columnId}": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "orgs",
          "in": "path",
          "required": true
        },
        {
          "schema": {
            "type": "string"
          },
          "name": "projectName",
          "in": "path",
          "required": true
        },
        {
          "schema": {
            "type": "string"
          },
          "name": "tableName",
          "in": "path",
          "required": true
        },
        {
          "schema": {
            "type": "string"
          },
          "name": "viewName",
          "in": "path",
          "required": true
        },
        {
          "schema": {
            "type": "string"
          },
          "name": "columnId",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "summary": "Table Group by Column",
        "operationId": "db-view-row-grouped-data-list",
        "description": "",
        "tags": [
          "DB view row"
        ],
        "parameters": [
          {
            "schema": {
              "type": "array"
            },
            "in": "query",
            "name": "fields"
          },
          {
            "schema": {
              "type": "array"
            },
            "in": "query",
            "name": "sort"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "where"
          },
          {
            "schema": {},
            "in": "query",
            "name": "nested",
            "description": "Query params for nested data"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/v1/db/data/{orgs}/{projectName}/{tableName}/group/{columnId}": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "orgs",
          "in": "path",
          "required": true
        },
        {
          "schema": {
            "type": "string"
          },
          "name": "projectName",
          "in": "path",
          "required": true
        },
        {
          "schema": {
            "type": "string"
          },
          "name": "tableName",
          "in": "path",
          "required": true
        },
        {
          "schema": {
            "type": "string"
          },
          "name": "columnId",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "summary": "Table Group by Column",
        "operationId": "db-table-row-grouped-data-list",
        "description": "",
        "tags": [
          "DB table row"
        ],
        "parameters": [
          {
            "schema": {
              "type": "array"
            },
            "in": "query",
            "name": "fields"
          },
          {
            "schema": {
              "type": "array"
            },
            "in": "query",
            "name": "sort"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "where"
          },
          {
            "schema": {},
            "in": "query",
            "name": "nested",
            "description": "Query params for nested data"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/v1/db/data/{orgs}/{projectName}/{tableName}/views/{viewName}": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "orgs",
          "in": "path",
          "required": true
        },
        {
          "schema": {
            "type": "string"
          },
          "name": "projectName",
          "in": "path",
          "required": true
        },
        {
          "schema": {
            "type": "string"
          },
          "name": "tableName",
          "in": "path",
          "required": true
        },
        {
          "schema": {
            "type": "string"
          },
          "name": "viewName",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "summary": "Table view row list",
        "operationId": "db-view-row-list",
        "description": "",
        "tags": [
          "DB view row"
        ],
        "parameters": [
          {
            "schema": {
              "type": "array"
            },
            "in": "query",
            "name": "fields"
          },
          {
            "schema": {
              "type": "array"
            },
            "in": "query",
            "name": "sort"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "where"
          },
          {
            "schema": {},
            "in": "query",
            "name": "nested",
            "description": "Query params for nested data"
          },
          {
            "schema": {
              "type": "number"
            },
            "in": "query",
            "name": "offset"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "summary": "Table view row create",
        "operationId": "db-view-row-create",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "tags": [
          "DB view row"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          }
        }
      }
    },
    "/api/v1/db/data/{orgs}/{projectName}/{tableName}/views/{viewName}/find-one": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "orgs",
          "in": "path",
          "required": true
        },
        {
          "schema": {
            "type": "string"
          },
          "name": "projectName",
          "in": "path",
          "required": true
        },
        {
          "schema": {
            "type": "string"
          },
          "name": "tableName",
          "in": "path",
          "required": true
        },
        {
          "schema": {
            "type": "string"
          },
          "name": "viewName",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "summary": "Table view row FindOne",
        "operationId": "db-view-row-find-one",
        "description": "",
        "tags": [
          "DB view row"
        ],
        "parameters": [
          {
            "schema": {
              "type": "array"
            },
            "in": "query",
            "name": "fields"
          },
          {
            "schema": {
              "type": "array"
            },
            "in": "query",
            "name": "sort"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "where"
          },
          {
            "schema": {},
            "in": "query",
            "name": "nested",
            "description": "Query params for nested data"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/v1/db/data/{orgs}/{projectName}/{tableName}/views/{viewName}/groupby": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "orgs",
          "in": "path",
          "required": true
        },
        {
          "schema": {
            "type": "string"
          },
          "name": "projectName",
          "in": "path",
          "required": true
        },
        {
          "schema": {
            "type": "string"
          },
          "name": "tableName",
          "in": "path",
          "required": true
        },
        {
          "schema": {
            "type": "string"
          },
          "name": "viewName",
          "in": "path",
          "required": true
        },
        {
          "schema": {
            "type": "string"
          },
          "in": "query",
          "name": "column_name",
          "description": "Column name of the column you want to group by, eg. `column_name=column1`"
        }
      ],
      "get": {
        "summary": "Table view row Group by",
        "operationId": "db-view-row-group-by",
        "description": "",
        "tags": [
          "DB view row"
        ],
        "parameters": [
          {
            "schema": {
              "type": "array"
            },
            "in": "query",
            "name": "sort"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "where"
          },
          {
            "schema": {
              "type": "number"
            },
            "in": "query",
            "name": "limit"
          },
          {
            "schema": {
              "type": "number"
            },
            "in": "query",
            "name": "offset"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/v1/db/data/{orgs}/{projectName}/{tableName}/views/{viewName}/count": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "orgs",
          "in": "path",
          "required": true
        },
        {
          "schema": {
            "type": "string"
          },
          "name": "projectName",
          "in": "path",
          "required": true
        },
        {
          "schema": {
            "type": "string"
          },
          "name": "tableName",
          "in": "path",
          "required": true
        },
        {
          "schema": {
            "type": "string"
          },
          "name": "viewName",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "summary": "Table view rows count",
        "operationId": "db-view-row-count",
        "description": "",
        "tags": [
          "DB view row"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "where"
          },
          {
            "schema": {},
            "in": "query",
            "name": "nested",
            "description": "Query params for nested data"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/v1/db/data/{orgs}/{projectName}/{tableName}/views/{viewName}/{rowId}": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "orgs",
          "in": "path",
          "required": true
        },
        {
          "schema": {
            "type": "string"
          },
          "name": "projectName",
          "in": "path",
          "required": true
        },
        {
          "schema": {
            "type": "string"
          },
          "name": "tableName",
          "in": "path",
          "required": true
        },
        {
          "schema": {
            "type": "string"
          },
          "name": "viewName",
          "in": "path",
          "required": true
        },
        {
          "schema": {
            "type": "string"
          },
          "name": "rowId",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "summary": "Table view row read",
        "operationId": "db-view-row-read",
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "description": "",
        "tags": [
          "DB view row"
        ]
      },
      "patch": {
        "summary": "Table view row update",
        "operationId": "db-view-row-update",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "tags": [
          "DB view row"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          }
        }
      },
      "delete": {
        "summary": "Table view row delete",
        "operationId": "db-view-row-delete",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "DB view row"
        ],
        "description": ""
      }
    },
    "/api/v1/db/data/{orgs}/{projectName}/{tableName}/views/{viewName}/{rowId}/exist": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "orgs",
          "in": "path",
          "required": true
        },
        {
          "schema": {
            "type": "string"
          },
          "name": "projectName",
          "in": "path",
          "required": true
        },
        {
          "schema": {
            "type": "string"
          },
          "name": "tableName",
          "in": "path",
          "required": true
        },
        {
          "schema": {
            "type": "string"
          },
          "name": "viewName",
          "in": "path",
          "required": true
        },
        {
          "schema": {
            "type": "string"
          },
          "name": "rowId",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "summary": "Table view row exist",
        "operationId": "db-view-row-exist",
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "description": "check row with provided primary key exists or not",
        "tags": [
          "DB view row"
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "xc-auth": {
        "name": "API Key",
        "type": "apiKey",
        "in": "header",
        "description": "JWT access token"
      },
      "xc-shared-base-id": {
        "name": "API Key",
        "type": "apiKey",
        "in": "header",
        "description": "Shared base uuid"
      }
    }
  }
}